rename sufix component
