+++ lib/Scenes/Page/HomePage/home_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:timeboxing/Scenes/Page/HomePage/Component/greeting_information.dart';
 import 'package:timeboxing/Scenes/Page/HomePage/Component/recomendation_component.dart';
-import 'package:timeboxing/Scenes/Page/HomePage/Component/tutorial.dart';
+import 'package:timeboxing/Scenes/Page/HomePage/Component/tutorial_component.dart';
 import 'package:timeboxing/Scenes/Page/HomePage/Component/weekly_date_picker_component.dart';
 import 'package:timeboxing/Scenes/Page/HomePage/Component/invitation_card.dart';
 import 'package:timeboxing/Scenes/Page/HomePage/Component/today_task.dart';
